+++ internal/database/parse_test.go
@@ -42,6 +42,6 @@ package database_test
 // 	}
 // }
 
-func ParseConfig(miscFile, s string) {
-	panic("unimplemented")
-}
+// func ParseConfig(miscFile, s string) {
+// 	panic("unimplemented")
+// }