+++ internal/actions/klone.go
@@ -2,11 +2,13 @@ package actions
 
 import (
 	"fmt"
+	"io"
 	"os"
 
-	"github.com/easbarba/onur/internal/common"
 	"github.com/go-git/go-git/v5"
 	"github.com/go-git/go-git/v5/plumbing"
+
+	"gitlab.com/easbarba/onur/internal/common"
 )
 
 // clone repository if none is found at folder
@@ -17,8 +19,9 @@ func Klone(folder, name, url, branch string) {
 	_, err := git.PlainClone(folder, false, &git.CloneOptions{
 		URL:           url,
 		ReferenceName: plumbing.ReferenceName(branchHead),
-		SingleBranch:  settings.SingleBranch,
-		Depth:         settings.Depth,
+		SingleBranch:  settings.Git.SingleBranch,
+		Depth:         settings.Git.Depth,
+		Progress:      io.Discard,
 	})
 
 	if err != nil {