+++ internal/storage/handlers.go
@@ -12,7 +12,8 @@
 *  You should have received a copy of the GNU General Public License
 *  along with Onur. If not, see <https://www.gnu.org/licenses/>.
  */
-package database
+
+package storage
 
 // import (
 // 	"encoding/json"